Completed Update
With this, the Website is officially complete. I'll wait for the info to translate into the actual website and verify the activation email to show the website completely. After this. No more code should be added unless I need to tweak some changes and updates on new projects later down the line, or making a version 3.0 later down the line. A good month was spent on this. Let's make sure the next project get's as much love an attention as this!